@@ -1,18 +1,25 @@
 {
     "name": "rndwiga/toolbox",
     "type": "package",
-    "description": "A package contains helper methods and classes that can be used in any project",
-    "keywords": ["ui", "rndwiga",  "admin","laravel"],
+    "description": "A package containing helper methods and classes that can be used in any project",
+    "keywords": ["toolbox", "rndwiga", "helpers", "utilities"],
     "authors": [
         {
             "name": "Raphael Ndwiga",
-            "email": "raphndwi@gmail.com"
+            "email": "raphael@raphaelndwiga.africa"
         }
     ],
     "require": {
-        "php": ">=7.1",
-        "monolog/monolog": "^1.24",
-        "vlucas/phpdotenv": "^3.4"
+        "php": "^8.1",
+        "monolog/monolog": "^2.0",
+        "vlucas/phpdotenv": "^5.0",
+        "ext-zlib": "*"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^9.0"
+    },
+    "suggest": {
+        "illuminate/support": "Required for Laravel integration"
     },
     "autoload": {
         "psr-4": {
@@ -22,8 +29,14 @@
             "src/Helpers/toolbox.php"
         ]
     },
+    "autoload-dev": {
+        "psr-4": {
+            "Rndwiga\\Toolbox\\Tests\\": "tests/"
+        }
+    },
+    "scripts": {
+        "test": "phpunit"
+    },
     "license": "MIT",
     "minimum-stability": "stable"
-
 }
-
