@@ -6,11 +6,11 @@
     "authors": [
         {
             "name": "Raphael Ndwiga",
-            "email": "raphndwi@gmail.com"
+            "email": "raphael@raphaelndwiga.africa"
         }
     ],
     "require": {
-        "php": ">=7.2",
+        "php": "^8.1",
         "rndwiga/mpesa": "@dev",
         "rndwiga/toolbox": "@dev",
         "monolog/monolog": "^1.24",
@@ -19,7 +19,8 @@
         "ramsey/uuid": "^3.8",
         "ext-json": "*",
         "ext-curl": "*",
-        "ext-openssl": "*"
+        "ext-openssl": "*",
+      "ext-zlib": "*"
     },
     "license": "MIT",
     "minimum-stability": "dev",
@@ -40,6 +41,12 @@
         }
     ],
     "require-dev": {
-
+        "pestphp/pest": "3.x-dev",
+        "mockery/mockery": "2.0.x-dev"
+    },
+    "config": {
+        "allow-plugins": {
+            "pestphp/pest-plugin": true
+        }
     }
 }
