@@ -7,17 +7,18 @@
     "authors": [
         {
             "name": "Raphael Ndwiga",
-            "email": "raphaelndwiga.packages@gmail.com"
+            "email": "raphael@raphaelndwiga.africa"
         }
     ],
     "require": {
-        "php": ">=7.1",
+        "php": "^8.1",
         "phpseclib/phpseclib": "^2.0",
       "ext-openssl": "*",
       "ext-json": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.5"
+        "pestphp/pest": "3.x-dev",
+        "mockery/mockery": "2.0.x-dev"
     },
     "autoload": {
         "psr-4": {
@@ -25,7 +26,10 @@
         }
     },
     "license": "MIT",
-    "minimum-stability": "stable"
-
+    "minimum-stability": "stable",
+    "config": {
+        "allow-plugins": {
+            "pestphp/pest-plugin": true
+        }
+    }
 }
-
