@@ -15,77 +15,62 @@ use Rndwiga\Mpesa\Libraries\MpesaApiConnection;
 
 class MpesaB2BCalls extends BaseRequest
 {
-
-    public function sendFundsToBusiness(){
-
+    /**
+     * Sample method to send funds to a business
+     * 
+     * @return string The API response
+     */
+    public function sendFundsToBusiness()
+    {
         $response = $this->setPartyB("600000")
-            ->setCommandID("BusinessToBusinessTransfer")
+            ->setCommandId("BusinessToBusinessTransfer")
             ->setSenderIdentifierType("4")
-            ->setReiver("4")
+            ->setReceiverIdentifierType(4)
             ->setAccountReference("BusinessPaybill")
             ->setRemarks("Business Payment To Business")
-            ->b2b(100);
+            ->makeB2bCall(100);
         return $response;
     }
 
-    public function makeB2bCall( $Amount){
-
-        if ($this->ApplicationStatus == 'live'){
-            $live = true;
-        }elseif ($this->ApplicationStatus == 'sandbox'){
-            $live = false;
-        }
-        else{
-            $live = null;
-        }
-
-        if(is_null($live)){
-            die("please declare the application status as defined in the documentation");
-        }
-
-        if( $live == true){
-            $url = 'https://api.safaricom.co.ke/mpesa/b2b/v1/paymentrequest';
-        }elseif ($live== false){
-            $url = 'https://sandbox.safaricom.co.ke/mpesa/b2b/v1/paymentrequest';
-        }else{
-            return json_encode(["Message"=>"invalid application status"]);
+    /**
+     * Make a B2B API call
+     * 
+     * @param int $amount The amount to transfer
+     * @param bool $verifySSL Whether to verify SSL certificates
+     * @return string The API response
+     */
+    public function makeB2bCall($amount, $verifySSL = true)
+    {
+        // Set the amount if it's provided as a parameter
+        if ($amount !== null) {
+            $this->setAmount($amount);
         }
 
-        $token = $this->generateAccessToken($live,$this->ConsumerKey,$this->ConsumerSecret);
+        // Validate required fields
+        if (!isset($this->InitiatorName) || !isset($this->SecurityCredential) || 
+            !isset($this->CommandID) || !isset($this->Amount) || 
+            !isset($this->PartyA) || !isset($this->PartyB) || 
+            !isset($this->QueueTimeOutURL) || !isset($this->ResultURL)) {
+            throw new \InvalidArgumentException("Missing required parameters for B2B call");
+        }
 
-        $curl = curl_init();
-        curl_setopt($curl, CURLOPT_URL, $url);
-        curl_setopt($curl, CURLOPT_HTTPHEADER, array('Content-Type:application/json','Authorization:Bearer '.$token)); //setting custom header
-        $curl_post_data = array(
+        // Prepare request data
+        $requestData = [
             'Initiator' => $this->InitiatorName,
             'SecurityCredential' => $this->SecurityCredential,
             'CommandID' => $this->CommandID,
             'SenderIdentifierType' => $this->SenderIdentifierType,
             'RecieverIdentifierType' => $this->ReceiverIdentifierType,
-            'Amount' => $Amount,
+            'Amount' => $this->Amount,
             'PartyA' => $this->PartyA,
             'PartyB' => $this->PartyB,
             'AccountReference' => $this->AccountReference,
             'Remarks' => $this->Remarks,
             'QueueTimeOutURL' => $this->QueueTimeOutURL,
             'ResultURL' => $this->ResultURL
-        );
-        $data_string = json_encode($curl_post_data);
-        curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
-        curl_setopt($curl, CURLOPT_POST, true);
-        curl_setopt($curl, CURLOPT_POSTFIELDS, $data_string);
+        ];
 
-        $curl_response = curl_exec($curl);
-        $err = curl_error($curl);
-
-        curl_close($curl);
-
-        if ($err) {
-            echo "cURL Error #:" . $err;
-            $curl_response = "cURL Error #:" . $err;
-            return $curl_response;
-        } else {
-            return $curl_response;
-        }
+        // Make the API request
+        return $this->makeApiRequest(self::B2B_ENDPOINT, $requestData, $verifySSL);
     }
 }
