@@ -1,10 +1,11 @@
 <?php
-
 /**
- * Created by PhpStorm.
- * User: rndwiga
- * Date: 3/27/18
- * Time: 3:29 PM
+ * AppLogger Class
+ *
+ * A class for logging application data using Monolog.
+ *
+ * @package Rndwiga\Toolbox\Infrastructure\Services
+ * @author Raphael Ndwiga <raphndwi@gmail.com>
  */
 namespace Rndwiga\Toolbox\Infrastructure\Services;
 
@@ -14,20 +15,48 @@ use Monolog\Logger;
 
 class AppLogger
 {
+    /**
+     * The name of the log file
+     *
+     * @var string
+     */
     private $fileName;
-    private $maxNumberOfLines = 10000;
-    private $logLevel;
 
+    /**
+     * The maximum number of lines to keep in the log file
+     *
+     * @var int
+     */
+    private $maxNumberOfLines = 10000;
+
+    /**
+     * The log level
+     *
+     * @var string
+     */
+    private $logLevel = 'debug';
+
+    /**
+     * The Monolog logger instance
+     *
+     * @var Logger
+     */
     private $logger;
 
+    /**
+     * The log storage path
+     *
+     * @var string
+     */
     private $logStorage;
 
     /**
      * AppLogger constructor.
-     * @param string $folderName
-     * @param string $fileName
+     * 
+     * @param string $folderName The folder name for storing logs
+     * @param string $fileName The name of the log file (without extension)
      */
-    public function __construct(string $folderName,string $fileName)
+    public function __construct(string $folderName, string $fileName)
     {
         $this->logger = $this->getLogger();
         $this->logStorage = (new AppStorage())->setLogFolder($folderName)->createStorage();
@@ -35,14 +64,9 @@ class AppLogger
     }
 
     /**
-     *This is a sample implementation of the logging class
-     */
-    public function sampleImplementation(){
-        (new AppLogger('testFolder','test_file'))->logInfo(['data']);
-    }
-
-    /**
-     * @return mixed
+     * Get the file name
+     * 
+     * @return string The file name
      */
     public function getFileName()
     {
@@ -50,7 +74,9 @@ class AppLogger
     }
 
     /**
-     * @param mixed $fileName
+     * Set the file name
+     * 
+     * @param string $fileName The file name
      * @return AppLogger
      */
     public function setFileName($fileName)
@@ -60,7 +86,9 @@ class AppLogger
     }
 
     /**
-     * @return mixed
+     * Get the maximum number of lines
+     * 
+     * @return int The maximum number of lines
      */
     public function getMaxNumberOfLines()
     {
@@ -68,7 +96,9 @@ class AppLogger
     }
 
     /**
-     * @param mixed $maxNumberOfLines
+     * Set the maximum number of lines
+     * 
+     * @param int $maxNumberOfLines The maximum number of lines
      * @return AppLogger
      */
     public function setMaxNumberOfLines($maxNumberOfLines)
@@ -78,72 +108,204 @@ class AppLogger
     }
 
     /**
-     * @return mixed
+     * Get the log level as a Monolog constant
+     * 
+     * @return int The log level constant
      */
     public function getLogLevel()
     {
-        switch ($this->logLevel){
-            case 'debug':
+        switch ($this->logLevel) {
+            case 'info':
+                return Logger::INFO;
+            case 'notice':
+                return Logger::NOTICE;
+            case 'warning':
+                return Logger::WARNING;
+            case 'error':
+                return Logger::ERROR;
+            case 'critical':
+                return Logger::CRITICAL;
+            case 'alert':
+                return Logger::ALERT;
+            case 'emergency':
+                return Logger::EMERGENCY;
+            default:
                 return Logger::DEBUG;
-                break;
         }
     }
 
     /**
-     * @param mixed $logLevel
+     * Set the log level
+     * 
+     * @param string $logLevel The log level (debug, info, notice, warning, error, critical, alert, emergency)
      * @return AppLogger
      */
     public function setLogLevel(string $logLevel)
     {
-        if (is_null($logLevel)){
+        $validLevels = ['debug', 'info', 'notice', 'warning', 'error', 'critical', 'alert', 'emergency'];
+
+        if (!in_array($logLevel, $validLevels)) {
             $logLevel = 'debug';
         }
+
         $this->logLevel = $logLevel;
         return $this;
     }
 
     /**
-     * @param string $loggerN
-     * @return Logger
+     * Get the logger instance
+     * 
+     * @param string $loggerName The name of the logger
+     * @return Logger The logger instance
      */
-    public function getLogger(string $loggerN = 'app.logger'): Logger
+    public function getLogger(string $loggerName = 'app.logger'): Logger
     {
-        return $this->logger = new Logger($loggerN);
+        return new Logger($loggerName);
     }
 
-    /***
-     * @param $dataToLog - json data to log
+    /**
+     * Log data at the specified level
+     * 
+     * @param array $dataToLog The data to log
+     * @param string $level The log level (debug, info, notice, warning, error, critical, alert, emergency)
+     * @return bool True if logging was successful, false otherwise
      */
-    public function logInfo(array $dataToLog){
+    public function log(array $dataToLog, string $level = 'debug')
+    {
+        $this->setLogLevel($level);
+
         // Trim log file to a max length
         $path = storagePath("{$this->logStorage}/{$this->getFileName()}.log");
-        if (! file_exists($path)) {
-            fopen($path, "w");
-        }
-        $lines = file($path);
-        if (count($lines) >= $this->getMaxNumberOfLines()) {
-            file_put_contents($path, implode('', array_slice($lines, -($this->getMaxNumberOfLines()), $this->getMaxNumberOfLines())));
-        }
-        // Define custom Monolog handler
-        try {
-            $handler = new StreamHandler($path, $this->getLogLevel());
-        } catch (\Exception $e) {
-        } //This will have both DEBUG and ERROR messages
-        $handler->setFormatter(new LineFormatter(null, null, true, true));
 
-        // Set defined handler and log the message
-        $this->logger->setHandlers([$handler]);
-        $this->logger->addDebug(json_encode($dataToLog)); //addError, addInfo, addWarning --TODO::enhance this
+        try {
+            if (!file_exists($path)) {
+                $file = fopen($path, "w");
+                if ($file === false) {
+                    return false;
+                }
+                fclose($file);
+            }
+
+            $lines = file($path);
+            if ($lines !== false && count($lines) >= $this->getMaxNumberOfLines()) {
+                $result = file_put_contents($path, implode('', array_slice($lines, -($this->getMaxNumberOfLines()), $this->getMaxNumberOfLines())));
+                if ($result === false) {
+                    return false;
+                }
+            }
+
+            // Define custom Monolog handler
+            $handler = new StreamHandler($path, $this->getLogLevel());
+            $handler->setFormatter(new LineFormatter(null, null, true, true));
+
+            // Set defined handler and log the message
+            $this->logger->setHandlers([$handler]);
+
+            // Log the message at the appropriate level
+            switch ($level) {
+                case 'info':
+                    $this->logger->info(json_encode($dataToLog));
+                    break;
+                case 'notice':
+                    $this->logger->notice(json_encode($dataToLog));
+                    break;
+                case 'warning':
+                    $this->logger->warning(json_encode($dataToLog));
+                    break;
+                case 'error':
+                    $this->logger->error(json_encode($dataToLog));
+                    break;
+                case 'critical':
+                    $this->logger->critical(json_encode($dataToLog));
+                    break;
+                case 'alert':
+                    $this->logger->alert(json_encode($dataToLog));
+                    break;
+                case 'emergency':
+                    $this->logger->emergency(json_encode($dataToLog));
+                    break;
+                default:
+                    $this->logger->debug(json_encode($dataToLog));
+            }
+
+            return true;
+        } catch (\Exception $e) {
+            return false;
+        }
     }
 
-    public function getLogFile(){
-        $path = storagePath("{$this->logStorage}/{$this->getFileName()}.json");
-        if (! file_exists($path)) {
-            fopen($path, "w");
-        }
-        $lines = file($path);
-        if (count($lines) == 0) {
-            file_put_contents($path, json_encode([]));
+    /**
+     * Log data at the debug level
+     * 
+     * @param array $dataToLog The data to log
+     * @return bool True if logging was successful, false otherwise
+     */
+    public function logDebug(array $dataToLog)
+    {
+        return $this->log($dataToLog, 'debug');
+    }
+
+    /**
+     * Log data at the info level
+     * 
+     * @param array $dataToLog The data to log
+     * @return bool True if logging was successful, false otherwise
+     */
+    public function logInfo(array $dataToLog)
+    {
+        return $this->log($dataToLog, 'info');
+    }
+
+    /**
+     * Log data at the warning level
+     * 
+     * @param array $dataToLog The data to log
+     * @return bool True if logging was successful, false otherwise
+     */
+    public function logWarning(array $dataToLog)
+    {
+        return $this->log($dataToLog, 'warning');
+    }
+
+    /**
+     * Log data at the error level
+     * 
+     * @param array $dataToLog The data to log
+     * @return bool True if logging was successful, false otherwise
+     */
+    public function logError(array $dataToLog)
+    {
+        return $this->log($dataToLog, 'error');
+    }
+
+    /**
+     * Get the log file path
+     * 
+     * @param string $extension The file extension (default: json)
+     * @return string The log file path
+     */
+    public function getLogFile(string $extension = 'json')
+    {
+        $path = storagePath("{$this->logStorage}/{$this->getFileName()}.{$extension}");
+
+        try {
+            if (!file_exists($path)) {
+                $file = fopen($path, "w");
+                if ($file !== false) {
+                    fclose($file);
+
+                    if ($extension === 'json') {
+                        file_put_contents($path, json_encode([]));
+                    }
+                }
+            } else if ($extension === 'json') {
+                $lines = file($path);
+                if ($lines !== false && count($lines) == 0) {
+                    file_put_contents($path, json_encode([]));
+                }
+            }
+        } catch (\Exception $e) {
+            // Silently fail and return the path anyway
         }
 
         return $path;
